+++ themes/helium/wordpress/search.php
@@ -10,6 +10,8 @@
 
 defined('ABSPATH') or die;
 
+use Timber\Timber;
+
 /*
  * Search results page
  */