+++ themes/helium/wordpress/searchform.php
@@ -10,6 +10,8 @@
 
 defined('ABSPATH') or die;
 
+use Timber\Timber;
+
 /*
  * The template for displaying Search Form in Search widget
  */